Only remove .gclient if it exists

+++ setup.py
@@ -71,7 +71,8 @@ def get_v8():
 
     if not os.path.isdir('v8/.git'):
         print 'downloading v8'
-        os.remove('.gclient')
+        if os.path.exists('.gclient'):
+            os.remove('.gclient')
         run('fetch v8')
     else:
         print 'updating v8'
